requirements.txt

fastapi
uvicorn
openai
PyPDF2
faiss-cpu
numpy
python-multipart
streamlit
requests
